requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
black==25.1.0
blinker==1.9.0
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
filelock==3.18.0
flake8==7.2.0
Flask==3.1.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kombu==5.5.3
llvmlite==0.44.0
MarkupSafe==3.0.2
mccabe==0.7.0
more-itertools==10.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.76.0
openai-whisper==20240930
opencv-python-headless==4.11.0.86
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
pytube==15.0.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yt-dlp==2025.3.31
